Refactoring UIStoresController
- Use file-scoped namespace
- Use primary constructor
- Extract Rates, Settings and Token actions into partial class
- Move private methods to partial classes
- Simple syntax cleanups
