+++ BTCPayServer/Controllers/UIStoresController.Users.cs
@@ -38,135 +38,134 @@ using NBitcoin;
 using NBitcoin.DataEncoders;
 using StoreData = BTCPayServer.Data.StoreData;
 
-namespace BTCPayServer.Controllers
+namespace BTCPayServer.Controllers;
+
+public partial class UIStoresController
 {
-    public partial class UIStoresController
+    [HttpGet("{storeId}/users")]
+    public async Task<IActionResult> StoreUsers()
     {
-        [HttpGet("{storeId}/users")]
-        public async Task<IActionResult> StoreUsers()
+        var vm = new StoreUsersViewModel { Role = StoreRoleId.Employee.Role };
+        await FillUsers(vm);
+        return View(vm);
+    }
+
+    [HttpPost("{storeId}/users")]
+    [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
+    public async Task<IActionResult> StoreUsers(string storeId, StoreUsersViewModel vm)
+    {
+        await FillUsers(vm);
+        if (!ModelState.IsValid)
         {
-            var vm = new StoreUsersViewModel { Role = StoreRoleId.Employee.Role };
-            await FillUsers(vm);
             return View(vm);
         }
 
-        [HttpPost("{storeId}/users")]
-        [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
-        public async Task<IActionResult> StoreUsers(string storeId, StoreUsersViewModel vm)
+        var roles = await repo.GetStoreRoles(CurrentStore.Id);
+        if (roles.All(role => role.Id != vm.Role))
         {
-            await FillUsers(vm);
-            if (!ModelState.IsValid)
-            {
-                return View(vm);
-            }
-
-            var roles = await _Repo.GetStoreRoles(CurrentStore.Id);
-            if (roles.All(role => role.Id != vm.Role))
-            {
-                ModelState.AddModelError(nameof(vm.Role), "Invalid role");
-                return View(vm);
-            }
+            ModelState.AddModelError(nameof(vm.Role), "Invalid role");
+            return View(vm);
+        }
             
-            var user = await _UserManager.FindByEmailAsync(vm.Email);
-            var isExistingUser = user is not null;
-            var isExistingStoreUser = isExistingUser && await _Repo.GetStoreUser(storeId, user!.Id) is not null;
-            var successInfo = string.Empty;
-            if (user == null)
+        var user = await userManager.FindByEmailAsync(vm.Email);
+        var isExistingUser = user is not null;
+        var isExistingStoreUser = isExistingUser && await repo.GetStoreUser(storeId, user!.Id) is not null;
+        var successInfo = string.Empty;
+        if (user == null)
+        {
+            user = new ApplicationUser
             {
-                user = new ApplicationUser
-                {
-                    UserName = vm.Email,
-                    Email = vm.Email,
-                    RequiresEmailConfirmation = _policiesSettings.RequiresConfirmedEmail,
-                    RequiresApproval = _policiesSettings.RequiresUserApproval,
-                    Created = DateTimeOffset.UtcNow
-                };
+                UserName = vm.Email,
+                Email = vm.Email,
+                RequiresEmailConfirmation = policiesSettings.RequiresConfirmedEmail,
+                RequiresApproval = policiesSettings.RequiresUserApproval,
+                Created = DateTimeOffset.UtcNow
+            };
 
-                var result = await _UserManager.CreateAsync(user);
-                if (result.Succeeded)
-                {
-                    var tcs = new TaskCompletionSource<Uri>();
-                    var currentUser = await _UserManager.GetUserAsync(HttpContext.User);
-
-                    _eventAggregator.Publish(new UserRegisteredEvent
-                    {
-                        RequestUri = Request.GetAbsoluteRootUri(),
-                        Kind = UserRegisteredEventKind.Invite,
-                        User = user,
-                        InvitedByUser = currentUser,
-                        CallbackUrlGenerated = tcs
-                    });
-                    
-                    var callbackUrl = await tcs.Task;
-                    var settings = await _settingsRepository.GetSettingAsync<EmailSettings>() ?? new EmailSettings();
-                    var info = settings.IsComplete()
-                        ? "An invitation email has been sent.<br/>You may alternatively"
-                        : "An invitation email has not been sent, because the server does not have an email server configured.<br/> You need to";
-                    successInfo = $"{info} share this link with them: <a class='alert-link' href='{callbackUrl}'>{callbackUrl}</a>";
-                }
-                else
-                {
-                    ModelState.AddModelError(nameof(vm.Email), "User could not be invited");
-                    return View(vm);
-                }
-            }
-
-            var roleId = await _Repo.ResolveStoreRoleId(storeId, vm.Role);
-            var action = isExistingUser
-                ? isExistingStoreUser ? "updated" : "added"
-                : "invited";
-            if (await _Repo.AddOrUpdateStoreUser(CurrentStore.Id, user.Id, roleId))
+            var result = await userManager.CreateAsync(user);
+            if (result.Succeeded)
             {
-                TempData.SetStatusMessageModel(new StatusMessageModel
+                var tcs = new TaskCompletionSource<Uri>();
+                var currentUser = await userManager.GetUserAsync(HttpContext.User);
+
+                eventAggregator.Publish(new UserRegisteredEvent
                 {
-                    Severity = StatusMessageModel.StatusSeverity.Success,
-                    AllowDismiss = false,
-                    Html = $"User {action} successfully." + (string.IsNullOrEmpty(successInfo) ? "" : $" {successInfo}")
+                    RequestUri = Request.GetAbsoluteRootUri(),
+                    Kind = UserRegisteredEventKind.Invite,
+                    User = user,
+                    InvitedByUser = currentUser,
+                    CallbackUrlGenerated = tcs
                 });
-                return RedirectToAction(nameof(StoreUsers));
+                    
+                var callbackUrl = await tcs.Task;
+                var settings = await settingsRepository.GetSettingAsync<EmailSettings>() ?? new EmailSettings();
+                var info = settings.IsComplete()
+                    ? "An invitation email has been sent.<br/>You may alternatively"
+                    : "An invitation email has not been sent, because the server does not have an email server configured.<br/> You need to";
+                successInfo = $"{info} share this link with them: <a class='alert-link' href='{callbackUrl}'>{callbackUrl}</a>";
             }
-
-            ModelState.AddModelError(nameof(vm.Email), $"The user could not be {action}");
-            return View(vm);
-        }
-
-        [HttpPost("{storeId}/users/{userId}")]
-        [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
-        public async Task<IActionResult> UpdateStoreUser(string storeId, string userId, StoreUsersViewModel.StoreUserViewModel vm)
-        {
-            var roleId = await _Repo.ResolveStoreRoleId(storeId, vm.Role);
-            var storeUsers = await _Repo.GetStoreUsers(storeId);
-            var user = storeUsers.First(user => user.Id == userId);
-            var isOwner = user.StoreRole.Id == StoreRoleId.Owner.Id;
-            var isLastOwner = isOwner && storeUsers.Count(u => u.StoreRole.Id == StoreRoleId.Owner.Id) == 1;
-            if (isLastOwner && roleId != StoreRoleId.Owner)
-                TempData[WellKnownTempData.ErrorMessage] = $"User {user.Email} is the last owner. Their role cannot be changed.";
-            else if (await _Repo.AddOrUpdateStoreUser(storeId, userId, roleId))
-                TempData[WellKnownTempData.SuccessMessage] = $"The role of {user.Email} has been changed to {vm.Role}.";
-            return RedirectToAction(nameof(StoreUsers), new { storeId, userId });
-        }
-
-        [HttpPost("{storeId}/users/{userId}/delete")]
-        [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
-        public async Task<IActionResult> DeleteStoreUser(string storeId, string userId)
-        {
-            if (await _Repo.RemoveStoreUser(storeId, userId))
-                TempData[WellKnownTempData.SuccessMessage] = "User removed successfully.";
             else
-                TempData[WellKnownTempData.ErrorMessage] = "Removing this user would result in the store having no owner.";
-            return RedirectToAction(nameof(StoreUsers), new { storeId, userId });
+            {
+                ModelState.AddModelError(nameof(vm.Email), "User could not be invited");
+                return View(vm);
+            }
         }
 
-        private async Task FillUsers(StoreUsersViewModel vm)
+        var roleId = await repo.ResolveStoreRoleId(storeId, vm.Role);
+        var action = isExistingUser
+            ? isExistingStoreUser ? "updated" : "added"
+            : "invited";
+        if (await repo.AddOrUpdateStoreUser(CurrentStore.Id, user.Id, roleId))
         {
-            var users = await _Repo.GetStoreUsers(CurrentStore.Id);
-            vm.StoreId = CurrentStore.Id;
-            vm.Users = users.Select(u => new StoreUsersViewModel.StoreUserViewModel()
+            TempData.SetStatusMessageModel(new StatusMessageModel
             {
-                Email = u.Email,
-                Id = u.Id,
-                Role = u.StoreRole.Role
-            }).ToList();
+                Severity = StatusMessageModel.StatusSeverity.Success,
+                AllowDismiss = false,
+                Html = $"User {action} successfully." + (string.IsNullOrEmpty(successInfo) ? "" : $" {successInfo}")
+            });
+            return RedirectToAction(nameof(StoreUsers));
         }
+
+        ModelState.AddModelError(nameof(vm.Email), $"The user could not be {action}");
+        return View(vm);
+    }
+
+    [HttpPost("{storeId}/users/{userId}")]
+    [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
+    public async Task<IActionResult> UpdateStoreUser(string storeId, string userId, StoreUsersViewModel.StoreUserViewModel vm)
+    {
+        var roleId = await repo.ResolveStoreRoleId(storeId, vm.Role);
+        var storeUsers = await repo.GetStoreUsers(storeId);
+        var user = storeUsers.First(user => user.Id == userId);
+        var isOwner = user.StoreRole.Id == StoreRoleId.Owner.Id;
+        var isLastOwner = isOwner && storeUsers.Count(u => u.StoreRole.Id == StoreRoleId.Owner.Id) == 1;
+        if (isLastOwner && roleId != StoreRoleId.Owner)
+            TempData[WellKnownTempData.ErrorMessage] = $"User {user.Email} is the last owner. Their role cannot be changed.";
+        else if (await repo.AddOrUpdateStoreUser(storeId, userId, roleId))
+            TempData[WellKnownTempData.SuccessMessage] = $"The role of {user.Email} has been changed to {vm.Role}.";
+        return RedirectToAction(nameof(StoreUsers), new { storeId, userId });
+    }
+
+    [HttpPost("{storeId}/users/{userId}/delete")]
+    [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
+    public async Task<IActionResult> DeleteStoreUser(string storeId, string userId)
+    {
+        if (await repo.RemoveStoreUser(storeId, userId))
+            TempData[WellKnownTempData.SuccessMessage] = "User removed successfully.";
+        else
+            TempData[WellKnownTempData.ErrorMessage] = "Removing this user would result in the store having no owner.";
+        return RedirectToAction(nameof(StoreUsers), new { storeId, userId });
+    }
+
+    private async Task FillUsers(StoreUsersViewModel vm)
+    {
+        var users = await repo.GetStoreUsers(CurrentStore.Id);
+        vm.StoreId = CurrentStore.Id;
+        vm.Users = users.Select(u => new StoreUsersViewModel.StoreUserViewModel()
+        {
+            Email = u.Email,
+            Id = u.Id,
+            Role = u.StoreRole.Role
+        }).ToList();
     }
 }
